fix(bank-security): change @Table annotation to reactive and change ID Long -> Integer.

+++ services/bank-security/src/main/java/am/banking/system/security/model/repository/UserTokenRepository.java
@@ -12,6 +12,6 @@ import java.util.Optional;
  * Time: 23:29:04
  */
 @Repository
-public interface UserTokenRepository extends ReactiveCrudRepository<UserToken, Long> {
+public interface UserTokenRepository extends ReactiveCrudRepository<UserToken, Integer> {
     Optional<UserToken> findByToken(String token);
 }
